refactor: remove ScreenControlHelper object
