--- app/src/main/java/com/bnyro/clock/util/ScreenControlHelper.kt
@@ -1,20 +1,0 @@
-package com.bnyro.clock.util
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.ui.platform.LocalView
-
-object ScreenControlHelper {
-    // https://stackoverflow.com/a/71293123/9652621
-    @Composable
-    fun KeepScreenOn() {
-        val currentView = LocalView.current
-        DisposableEffect(Unit) {
-            currentView.keepScreenOn = true
-
-            onDispose {
-                currentView.keepScreenOn = false
-            }
-        }
-    }
-}
